requirements.txt
asttokens==3.0.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dash==3.2.0
dash-iconify==0.1.2
dash-mantine-components==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
Flask==3.1.2
fonttools==4.60.0
graphviz==0.21
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
llvmlite==0.44.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mplfinance==0.12.10b0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pandas-ta==0.4.71b0
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
prompt_toolkit==3.0.52
psutil==7.1.0
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.1.0
requests==2.32.5
retrying==1.4.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
ttkthemes==3.2.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
